revert fitted face points (points in the center of surface quads for 27-node triquadratic hexahedra) to using a legacy relaxation method.
I was going to try to set up mid-face points similar to the 'non-legacy' method for edges but I realized this wouldn't work as easily, since you need to know the type of surface element (quad, triangle, hexagon, etc) in order to know how many verticies are on the face.
I'm open to thoughts/suggestions on other ways to do this but this method works for the time being.

modified:   vtkVmtk/Misc/vtkvmtkLinearToQuadraticMeshFilter.cxx