Added compression options for the vmtknumpywriter script for hdf5 datasets.
This module is tested and working. For a large uncompressed hdf5 dataset
(2.2GB), the compressed version was shrunk by over an order of magnitude
(roughly 148MB). gzip compression is hardcoded as a fixed parameter as it is the
most portable of the options, having support on every version of HDF5 (not only
on h5py instances). See
http://docs.h5py.org/en/latest/high/dataset.html#lossless-compression-filters
for more details on other compression filters.

Compression is not supported for pickle objects because it is a pain in the rear
to guarrentee string encoding error will not be thrown if a object is pickled on
a python 3 version of vmtk and opened on a python 2 version of vmtk (or vice
versa). see http://lucumr.pocoo.org/2014/1/5/unicode-in-2-and-3/ for more
details on working with zlib compression and python pickle objects.