Set cmake min version to 3.3
Minimum version was updated, and CMake Policy CMP0025 is set to NEW in
order to allow us to detect the features of custom compiler packages.
This is a requirement for using the anaconda c and cxx compilers, as
they implement a spec-compliant, but non standard compiler set on OSX
and Linux in order to standardize the compilers used to build non python
dependencies.

Also needed to set CMP0033 to OLD behavior. This allows for us to
continue using the cmake command export_library_depends(), which was
replaced in cmake 3.0+. Failure to set this behavior yields a fatal
error.