Modernize cmake export commands.
Removed dependency on setting CMP0022 and CMP0033 to the old behavior
(which is giving deprication notices on cmake 3.9+). To do this, the
cmake export_library_dependencies needed to be removed. Instead of
keeping this at the top level cmake lists file, it is replaced by an
install export command, which pulls from VMTK-Targets. The VMTK-Targets
exports are defined from install commands which process c++ targets.