remove vmtkimagereader InputDirectoryName option
This is an immediate removal of both the python and c++ code which
handled reading in dicom files specified from a directory name. This was
performed for a number of reason:

1) it is now best practice to use the ITKIO module for loading files.
This utility can be generally regarded as being more accurate (by
keeping translation and rotation matrices aligned).

2) incompatibility with VTK 8.1.0. With c++ 14 onwards and VTK 8.1.0+,
the forward declarations in the vtkDICOMImageReader.h were somehow
overriding the full definitions when we included the DICOMParser.h files
in our derived class.

3) Using a simple directory name can lead to ambiguity if there is more
than one dicom image series in a directory. This is handled
transparently by ITKIO by requiring that the exact file/series desired is
specified.