Fix image compare script for newer vtk versions.
Prior to VTK 6.0+, the image.SetScalarType() was used to populate
pipeline information with scalar meta-data. This function was removed in
favor of casting methods. However, As users should not be comparing
images with differing types, the code was changed to only check that
scalar types were the same between the input and reference images, and
raise an error otherwise.