vastly improved method of rendering and speed.
Off the start vtkPolyDataConnectivityFilter is ran only once to extract
and label all regions. From there it is called again to extract only the
regionId which was selected by the user. This is added to it's own
actor in the scene (rather than adding every component as
individual actors). Unlimited selections and 'undos' can be performed.