conda: Attempt to fix failure to find pytest adding it as a build dependency
This is an attempt to fix error like the following reported on linux, macOS
and Windows:

  +pytest ./build/tests/
  /Users/travis/miniconda/conda-bld/vmtk_1531865637554/test_tmp/run_test.sh: line 7: pytest: command not found

Worth noting that the recipe didn't change since the last successful build
from May 30 2018. See [1]

The issue with the current build is that it doesn't install pytest despite
the fact it is listed as a test requirement.

[1] https://ci.appveyor.com/project/rlizzo/vmtk-bh6uc/build/1.0.227/job/5kx5skmqxus6svl6#L2849