fix: update reference data for otsu tests (#392)
VMTK otsu tests are not very sophisticated, and there is some small
difference using ITK5 versus using ITK4 that causes the Otsu
tests to fail.

Tests take the difference between the filter output and the
reference image. If any pixel value is above a threshold the test
fails. The difference image has pixels that wrap around the type,
e.g. given uint8 image A filled with zeros and and unint8 image B
filled with ones the output of A-B will give 255 for each pixel. The
difference image range maximum is larger than the tolerance causing
the test to fail.

Solve the issue by adding different versions of test that
are skipped depending on version of ITK.

Add a class that provides an interface to the itkVersion
class. This includes adding a test that python wrapped functionality
is exposed.

chore: pep8 fixes