COMP: Fix build with latest VTK master
A backward-incompatible changes were introduced in public VTK API (such as IntervalInformation from a protected member
was renamed to publicly available vtkIntervalInformation type).

The change was not associated with a VTK version change (not even patch), therefore a compile check was added
to detect what VTK version is used and make VMTK build correctly with both the latest VTK-9.1.0 and earlier versions.